Simplify config ui
Remove environment config info, checklist. Move additional config item info into a popover
